Promote render::Job to render::Task::Job
