+++ libraries/render-utils/src/HitEffect.h
@@ -17,7 +17,7 @@ public:
     HitEffect();
     
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
-    typedef render::Job::Model<HitEffect> JobModel;
+    using JobModel = render::Task::Job::Model<HitEffect>;
     
     const gpu::PipelinePointer& getHitEffectPipeline();
