+++ libraries/render-utils/src/RenderDeferredTask.h
@@ -22,14 +22,14 @@ class SetupDeferred {
 public:
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
 
-    typedef render::Job::Model<SetupDeferred> JobModel;
+    using JobModel = render::Task::Job::Model<SetupDeferred>;
 };
 
 class PrepareDeferred {
 public:
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
 
-    typedef render::Job::Model<PrepareDeferred> JobModel;
+    using JobModel = render::Task::Job::Model<PrepareDeferred>;
 };
 
 
@@ -37,7 +37,7 @@ class RenderDeferred {
 public:
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
 
-    typedef render::Job::Model<RenderDeferred> JobModel;
+    using JobModel = render::Task::Job::Model<RenderDeferred>;
 };
 
 class ToneMappingDeferred {
@@ -46,21 +46,21 @@ public:
 
     ToneMappingEffect _toneMappingEffect;
 
-    typedef render::Job::Model<ToneMappingDeferred> JobModel;
+    using JobModel = render::Task::Job::Model<ToneMappingDeferred>;
 };
 
 class DrawOpaqueDeferred {
 public:
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext, const render::ItemIDsBounds& inItems);
 
-    typedef render::Job::ModelI<DrawOpaqueDeferred, render::ItemIDsBounds> JobModel;
+    using JobModel = render::Task::Job::ModelI<DrawOpaqueDeferred, render::ItemIDsBounds>;
 };
 
 class DrawTransparentDeferred {
 public:
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext, const render::ItemIDsBounds& inItems);
 
-    typedef render::Job::ModelI<DrawTransparentDeferred, render::ItemIDsBounds> JobModel;
+    using JobModel = render::Task::Job::ModelI<DrawTransparentDeferred, render::ItemIDsBounds>;
 };
 
 class DrawStencilDeferred {
@@ -70,14 +70,14 @@ public:
 
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
 
-    typedef render::Job::Model<DrawStencilDeferred> JobModel;
+    using JobModel = render::Task::Job::Model<DrawStencilDeferred>;
 };
 
 class DrawBackgroundDeferred {
 public:
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
 
-    typedef render::Job::Model<DrawBackgroundDeferred> JobModel;
+    using JobModel = render::Task::Job::Model<DrawBackgroundDeferred>;
 };
 
 class DrawOverlay3D {
@@ -86,15 +86,15 @@ public:
     static const gpu::PipelinePointer& getOpaquePipeline();
     
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
-    
-    typedef render::Job::Model<DrawOverlay3D> JobModel;
+
+    using JobModel = render::Task::Job::Model<DrawOverlay3D>;
 };
 
 class Blit {
 public:
     void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
 
-    typedef render::Job::Model<Blit> JobModel;
+    using JobModel = render::Task::Job::Model<Blit>;
 };
 
 class RenderDeferredTask : public render::Task {
@@ -103,7 +103,7 @@ public:
     RenderDeferredTask();
     ~RenderDeferredTask();
 
-    render::Jobs _jobs;
+    render::Task::Jobs _jobs;
     
     int _drawDebugDeferredBufferIndex = -1;
     int _drawStatusJobIndex = -1;
