+++ libraries/render/src/render/DrawTask.h
@@ -9,216 +9,14 @@
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
 
-#ifndef hifi_render_Task_h
-#define hifi_render_Task_h
+#ifndef hifi_render_DrawTask_h
+#define hifi_render_DrawTask_h
 
 #include "Engine.h"
 #include "gpu/Batch.h"
-#include <PerfStat.h>
 
 
 namespace render {
-
-template <class T> void jobRun(T& jobModel, const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext) {
-    jobModel.run(sceneContext, renderContext);
-}
-template <class T, class I> void jobRunI(T& jobModel, const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, const I& input) {
-    jobModel.run(sceneContext, renderContext, input);
-}
-template <class T, class O> void jobRunO(T& jobModel, const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, O& output) {
-    jobModel.run(sceneContext, renderContext, output);
-}
-template <class T, class I, class O> void jobRunIO(T& jobModel, const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, const I& input, O& output) {
-    jobModel.run(sceneContext, renderContext, input, output);
-}
-
-class Job {
-public:
-
-    // Varying represent a varying piece of data
-    class Varying {
-    public:
-
-        Varying(const Varying& var): _concept(var._concept) {}
-
-        Varying() {}
-        template <class T> 
-        Varying(const T& data) : _concept(std::make_shared<Job::Varying::Model<T>>(data)) {}
-
-        // Access the _data contained win the concept explicitely
-        template <class T> T& edit() { return std::static_pointer_cast<Model<T>>(_concept)->_data; }
-        template <class T> const T& get() const { return std::static_pointer_cast<const Model<T>>(_concept)->_data; }
-
-    protected:
-        friend class Job;
-
-        std::vector<std::weak_ptr<Job>> _consumerJobs;
-
-        void addJobConsumer(const std::shared_ptr<Job>& job) {
-            _consumerJobs.push_back(job);
-        }
- 
-        class Concept {
-        public:
-            virtual ~Concept() = default;
-        };
-        template <class T> class Model : public Concept {
-        public:
-            typedef T Data;
-            Data _data;
-            Model(const Model& source): _data(source.data) {}
-            Model(const Data& data): _data(data) {}
-            virtual ~Model() {}
-        };
-
-        std::shared_ptr<Concept> _concept;
-    };
-
-    Job(const Job& other) : _concept(other._concept) {}
-    ~Job();
-
-    bool isEnabled() const { return _concept->isEnabled(); }
-    void setEnabled(bool isEnabled) { _concept->setEnabled(isEnabled); }
-
-    const std::string& getName() const { return _concept->getName(); }
-    const Varying getInput() const { return _concept->getInput(); }
-    const Varying getOutput() const { return _concept->getOutput(); }
-
-    template <class T> T& edit() {
-        auto theConcept = std::dynamic_pointer_cast<typename T::JobModel>(_concept);
-        assert(theConcept);
-        return theConcept->_data;
-    }
-    template <class T> const T& get() const {
-        auto theConcept = std::dynamic_pointer_cast<typename T::JobModel>(_concept);
-        assert(theConcept);
-        return theConcept->_data;
-    }
-
-    void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext) {
-        PerformanceTimer perfTimer(getName().c_str());
-        PROFILE_RANGE(getName().c_str());
-        _concept->run(sceneContext, renderContext);
-    }
-
-protected:
-public:
-
-    class Concept {
-        std::string _name;
-        bool _isEnabled = true;
-    public:
-        Concept() : _name() {}
-        Concept(const std::string& name) : _name(name) {}
-        virtual ~Concept() = default;
-        
-        void setName(const std::string& name) { _name = name; }
-        const std::string& getName() const { return _name; }
-
-        bool isEnabled() const { return _isEnabled; }
-        void setEnabled(bool isEnabled) { _isEnabled = isEnabled; }
-
-        virtual const Varying getInput() const { return Varying(); }
-        virtual const Varying getOutput() const { return Varying(); }
-        virtual void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext) = 0;
-    };
-
-    Job(Concept* concept) : _concept(concept) {}
-
-public:
-    template <class T> class Model : public Concept {
-    public:
-        typedef T Data;
-
-        Data _data;
-
-        Model() {}
-        Model(const std::string& name): Concept(name) {}
-        Model(Data data): _data(data) {}
-        Model(Data data, const std::string& name): Concept(name), _data(data) {}
-
-        void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext) {
-            if (isEnabled()) {
-                jobRun(_data, sceneContext, renderContext);
-            }
-        }
-    };
-
-    template <class T, class I> class ModelI : public Concept {
-    public:
-        typedef T Data;
-        typedef I Input;
-
-        Data _data;
-        Varying _input;
-
-        const Varying getInput() const { return _input; }
-
-        ModelI(const std::string& name, const Varying& input, Data data = Data()) : Concept(name), _data(data), _input(input) {}
-        ModelI(const std::string& name, Data data): Concept(name), _data(data) {}
-
-        void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext) {
-            if (isEnabled()) {
-                jobRunI(_data, sceneContext, renderContext, _input.get<I>());
-            }
-        }
-    };
-
-    template <class T, class O> class ModelO : public Concept {
-    public:
-        typedef T Data;
-        typedef O Output;
-
-        Data _data;
-        Varying _output;
-
-        const Varying getOutput() const { return _output; }
-
-        ModelO(const std::string& name): Concept(name), _output(Output()) {
-            
-        }
-
-        ModelO(const std::string& name, Data data): Concept(name), _data(data), _output(Output()) {}
-
-        void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext) {
-            if (isEnabled()) {
-                jobRunO(_data, sceneContext, renderContext, _output.edit<O>());
-            }
-        }
-    };
-
-    template <class T, class I, class O> class ModelIO : public Concept {
-    public:
-        typedef T Data;
-        typedef I Input;
-        typedef O Output;
-
-        Data _data;
-        Varying _input;
-        Varying _output;
-
-        const Varying getInput() const { return _input; }
-        const Varying getOutput() const { return _output; }
-
-        ModelIO(const std::string& name, const Varying& input, Data data = Data()): Concept(name), _data(data), _input(input), _output(Output()) {}
-        ModelIO(const std::string& name, Data data, Output output): Concept(name), _data(data), _output(output) {}
-
-        void setInput(const Varying& input) { _input = input; }
-
-        void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext) {
-            if (isEnabled()) {
-                jobRunIO(_data, sceneContext, renderContext, _input.get<I>(), _output.edit<O>());
-            }
-        }
-    };
-
-    std::shared_ptr<Concept> _concept;
-};
-
-
-
-
-typedef std::vector<Job> Jobs;
 
 void cullItems(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, const ItemIDsBounds& inItems, ItemIDsBounds& outITems);
 void depthSortItems(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, bool frontToBack, const ItemIDsBounds& inItems, ItemIDsBounds& outITems);
@@ -236,25 +34,25 @@ public:
 
     void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, ItemIDsBounds& outItems);
 
-    typedef Job::ModelO<FetchItems, ItemIDsBounds> JobModel;
+    using JobModel = Task::Job::ModelO<FetchItems, ItemIDsBounds>;
 };
 
 class CullItems {
 public:
     void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, const ItemIDsBounds& inItems, ItemIDsBounds& outItems);
-    typedef Job::ModelIO<CullItems, ItemIDsBounds, ItemIDsBounds> JobModel;
+    using JobModel = Task::Job::ModelIO<CullItems, ItemIDsBounds, ItemIDsBounds>;
 };
 
 class CullItemsOpaque {
 public:
     void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, const ItemIDsBounds& inItems, ItemIDsBounds& outItems);
-    typedef Job::ModelIO<CullItemsOpaque, ItemIDsBounds, ItemIDsBounds> JobModel;
+    using JobModel = Task::Job::ModelIO<CullItemsOpaque, ItemIDsBounds, ItemIDsBounds>;
 };
 
 class CullItemsTransparent {
 public:
     void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, const ItemIDsBounds& inItems, ItemIDsBounds& outItems);
-    typedef Job::ModelIO<CullItemsTransparent, ItemIDsBounds, ItemIDsBounds> JobModel;
+    using JobModel = Task::Job::ModelIO<CullItemsTransparent, ItemIDsBounds, ItemIDsBounds>;
 };
 
 class DepthSortItems {
@@ -265,14 +63,14 @@ public:
 
     void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext, const ItemIDsBounds& inItems, ItemIDsBounds& outITems);
 
-    typedef Job::ModelIO<DepthSortItems, ItemIDsBounds, ItemIDsBounds> JobModel;
+    using JobModel = Task::Job::ModelIO<DepthSortItems, ItemIDsBounds, ItemIDsBounds>;
 };
 
 class DrawLight {
 public:
     void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext);
 
-    typedef Job::Model<DrawLight> JobModel;
+    using JobModel = Task::Job::Model<DrawLight>;
 };
 
 class DrawBackground {
@@ -296,7 +94,6 @@ public:
 
 };
 
-
 // A map of ItemIDs allowing to create bucket lists of SHAPE type items which are filtered by their
 // Material 
 class ItemMaterialBucketMap : public std::map<model::MaterialFilter, ItemIDs, model::MaterialFilter::Less> {
@@ -313,4 +110,4 @@ void materialSortItems(const SceneContextPointer& sceneContext, const RenderCont
 
 }
 
-#endif // hifi_render_Task_h
+#endif // hifi_render_DrawTask_h
